% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{app_sys}
\usage{
app_sys(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
within your package. The default, none, returns the root of the app.}
}
\description{
Access files for the shiny app from the package installation directory.
}
\details{
# Note
If you manually change your package name in the `DESCRIPTION`,
don't forget to change it here too, and in the config file (`inst/config.yml`).

For a safer name change mechanism, use the `golem::set_golem_name()` function.
}
