% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_triangle_bundle.R
\name{create_triangle_bundle}
\alias{create_triangle_bundle}
\title{Create Triangle Bundle}
\usage{
create_triangle_bundle(
  loss_data,
  age_increment = 12,
  origin_col = "accident_year",
  age_col = "devt",
  value_col = "paid"
)
}
\arguments{
\item{loss_data}{initial aggregated loss data as a `data.frame`}

\item{age_increment}{increment in months between subsequent maturity periods}

\item{origin_col, age_col, value_col}{column names as strings}
}
\value{
list of class "triangle_bundle" with an added attribute describing
  which metric the bundle describes (i.e. paid, reported, counts, etc.)
}
\description{
Create a "bundle" of triangle related items from an input loss dataset. The function
returns a list of class "triangle_bundle" and an attribute describing which metric the
bundle describes (i.e. paid, reported, counts, etc.).

The resulting list contains:

  - aggregated data filtered for ages of maturity from the `age_increment` argument

  - triangle data derived off the aggregated data

  - the actual spread out triangle

  - age_to_age data

  - the age_to_age spread out triangle

  - averages (currently only straight and weighted)

  - initial selections for the LDF's and derived CDF's.
}
\examples{
library(lossdevtapp)
library(dplyr)

# create default paid triangle bundle
tri_paid_bundle <- create_triangle_bundle(loss_data_all)

# check out the structure
str(tri_paid_bundle)

# view the paid triangle
View(tri_paid_bundle$triangle)

# view the age-to-age triangle
View(tri_paid_bundle$age_to_age_triangle)

# derive a similar bundle for reported dollars and counts
tri_rept_bundle <- create_triangle_bundle(loss_data_all, value_col = "reported")
tri_cnts_bundle <- create_triangle_bundle(loss_data_all, value_col = "n_claims")
}
