% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{date_utils}
\alias{date_utils}
\alias{end_of_month}
\alias{beg_of_month}
\alias{start_of_month}
\title{Date Utility Functions}
\usage{
end_of_month(date)

beg_of_month(date)

start_of_month(date)
}
\arguments{
\item{date}{Character string or Date representing the date to manipulate.}
}
\value{
Returns the Start or End Date as a Date.
}
\description{
Date utility helpers for deriving start/end dates.
}
\examples{
# end_of_month ------------------------------------------------------------

# character input
end_of_month("2020-08-13")

# date input
end_of_month(as.Date("2020-08-13"))


# beg_of_month ------------------------------------------------------------

# character input
beg_of_month("2020-08-13")

# date input
beg_of_month(as.Date("2020-08-13"))


# start_of_month ----------------------------------------------------------

# character input
start_of_month("2020-08-13")

# date input
start_of_month(as.Date("2020-08-13"))

}
