% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derive_triangles.R
\name{derive_triangles}
\alias{derive_triangles}
\title{Derive Triangles}
\usage{
derive_triangles(
  loss_dat,
  type = c("paid", "reported", "case", "n_claims"),
  limit = NULL
)
}
\arguments{
\item{loss_dat}{loss data}

\item{type}{paid, reported, case, or n_claims}

\item{limit}{optinal limit}
}
\value{
list of triangle data
}
\description{
Derive Triangles
}
