% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dev_tri.R
\name{dev_tri}
\alias{dev_tri}
\title{Development Triangle Class}
\usage{
dev_tri(origin, age, value, value_label = NULL, latest_eval_date = NULL)
}
\arguments{
\item{origin, age, value}{columns necessary to generate a `dev_tri`}

\item{value_label}{optional label for the values (i.e. paid, incurred)}

\item{latest_eval_date}{optional - specify latest val date}
}
\description{
Development Triangle Class
}
\examples{
library(lossdevtapp)

my_triangle <- dev_tri(
  origin = loss_data$accident_year,
  age = loss_data$devt,
  value = loss_data$payment,
  value_label = "paid",
  latest_eval_date = max(loss_data$eval_date)
)

class(my_triangle)
str(my_triangle)
print(my_triangle)
view.dev_tri(my_triangle)

}
