% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{end_of_month}
\alias{end_of_month}
\title{End of Month}
\usage{
end_of_month(date)
}
\arguments{
\item{date}{character or date representation of a date.}
}
\value{
date
}
\description{
End of Month
}
\examples{
# character input
end_of_month("2020-08-13")
# date input
end_of_month(as.Date("2020-08-13"))
}
