% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback-helpers.R
\name{feedback}
\alias{feedback}
\alias{msg_field}
\alias{msg_value}
\alias{msg_done}
\alias{msg_bullet}
\alias{msg_err}
\alias{msg_path}
\alias{msg_info}
\alias{msg_code}
\alias{msg_feedback}
\title{Feedback Message Utility Functions}
\usage{
msg_field(x)

msg_value(x)

msg_done(x)

msg_bullet(x, bullet = cli::symbol$bullet)

msg_err(x)

msg_path(x)

msg_info(x)

msg_code(x)

msg_feedback(x)
}
\arguments{
\item{x}{The string passed to various `msg_` functions.}

\item{bullet}{What to use for the message's `bullet`. Defaults to `cli::symbol$bullet`}
}
\description{
A set of helper functions for providing verbose feedback to the developer using
this packages functions.
}
\seealso{
- [usethis::ui-questions()]
  - [cli::list_symbols()]

Other Feedback Utilities: 
\code{\link{indent}()},
\code{\link{inform}()}
}
\concept{Feedback Utilities}
\keyword{feedback}
