% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback-helpers.R
\name{indent}
\alias{indent}
\title{Indent}
\usage{
indent(x, first = "  ", indent = first)
}
\arguments{
\item{x}{The string passed to various `msg_` functions.}

\item{first}{what to indent with - defaults to `"  "`.}

\item{indent}{indentation of next line - defaults to `first`}
}
\value{
string
}
\description{
Indentation around various `msg_` feedback functions.
}
\seealso{
Other Feedback Utilities: 
\code{\link{feedback}},
\code{\link{inform}()}
}
\concept{Feedback Utilities}
