% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossdevtapp-package.R
\docType{package}
\name{lossdevtapp-package}
\alias{lossdevtapp}
\alias{lossdevtapp-package}
\title{lossdevtapp: Actuarial Loss Development}
\description{
Actuarial Loss Development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/lossdevtapp}
  \item Report bugs at \url{https://github.com/jimbrig/lossdevtapp/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com}

}
\keyword{internal}
