% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_header.R, R/app_ui.R
\name{app_header}
\alias{app_header}
\alias{shiny-ui}
\alias{app_ui}
\title{App Header}
\usage{
app_header(title = "Loss Development", ...)

app_ui(request, ...)
}
\arguments{
\item{title}{App title to be placed in the header, above the sidebar.}

\item{...}{
  Arguments passed on to \code{\link[shinydashboardPlus:dashboardHeader]{shinydashboardPlus::dashboardHeader}}
  \describe{
    \item{\code{titleWidth}}{The width of the title area. This must either be a number
which specifies the width in pixels, or a string that specifies the width
in CSS units.}
    \item{\code{disable}}{If \code{TRUE}, don't display the header bar.}
    \item{\code{.list}}{An optional list containing items to put in the header. Same as
the \code{...} arguments, but in list format. This can be useful when
working with programmatically generated items.}
    \item{\code{leftUi}}{Items that will appear on the left part of the navbar. Should
be wrapped in a tagList.}
    \item{\code{controlbarIcon}}{Customize the trigger icon of the right sidebar.}
    \item{\code{fixed}}{Whether the navbar is fixed-top or not. FALSE by default.}
  }}

\item{request}{Internal parameter for `{shiny}`.}
}
\value{
a [shinydashboard::dashboardHeader()]

The user interface definition, without modifications or side effects.
}
\description{
Shiny App Dashboard's Header Function

Functions that build the Shiny App's UI:

  - `app_ui`: Main UI function

  - `app_header`: wrapper around [shinydashboard::dashboardHeader()]

  - `app_sidebar`: wrapper around [shinydashboard::dashboardSidebar()]

  - `app_body`: wrapper around [shinydashboard::dashboardBody()]

Shiny App's User Interface Function.
}
\seealso{
[shiny::shinyUI], [shinydashboard::dashboardPage()]
}
