% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_claims.R
\name{simulate_claims}
\alias{simulate_claims}
\title{simulate_claims}
\usage{
simulate_claims(
  n_claims = 1000,
  start_date = "2015-01-01",
  end_date = Sys.Date(),
  seed = 12345,
  loss_distribution = "lnorm",
  params = list(mean_log = 7.5, sd_log = 1.5),
  status_prob_open = 0.96,
  cache = FALSE,
  ...
)
}
\arguments{
\item{n_claims}{Numeric - Number of claims to be simulated.}

\item{start_date, end_date}{Character/Date - Start and End dates for simulation to create claims within (experience_period).}

\item{seed}{Numeric - the seed is used to isolate randomness during statistical simulations.}

\item{loss_distribution}{Character - must be one of the distributions mentioned in the details below. Defaults to lognormal.}

\item{params}{Parameters associated with the specified `loss_distribution`
in a list (i.e. `list(mean_log = 7.5, sd_log = 1.5)` for lognormal distribution).}

\item{status_prob_open}{Numeric - must be within `0 < x < 1` and represents
probability a claim is open when running binomial simulations for claims'
status.}

\item{cache}{Boolean/Logical - enable caching?}

\item{...}{If needed}
}
\value{
The return value, if any, from executing the function.
}
\description{
A function to simulate *transactional* actuarial claims/loss
  data for Property Casualty Insurance.
}
\details{
Severity/Loss Distributions:
 - Normal: `norm` - Parameters are `mean` and `sd`.
 - Lognormal: `lnorm` - Parameters are `meanlog` and `sdlog`.
 - Gamma: `gamma` - Shape, Rate, Scale
 - LogGamma: `lgamma` - Shapelog, Ratelog
 - Pareto: `pareto` - Shape and Scale
 - Weibull: `weibull` - Shape and Scale
 - Generalized Beta: `genbeta` - Shape1, Shape2, Shape3, Rate, Scale
}
