% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_triangles.R
\name{triangles_module}
\alias{triangles_module}
\alias{mod_triangles_ui}
\alias{mod_triangles_server}
\title{Triangles Module}
\usage{
mod_triangles_ui(id, loss_data = loss_data_all)

mod_triangles_server(id, loss_data, selected_eval)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{loss_data}{loss data}

\item{selected_eval}{selected evaluation date}

\item{id}{ID associated with UI counterpart}
}
\description{
Shiny module containing `mod_triangles_ui` and `mod_triangles_server`, respectively.

This module renders a user interface for displaying and analyzing actuarial
loss data in the form of loss development triangles.

A shiny Module.
}
